requirements.txt
openai
langchain
python-dotenv
tqdm
langchain-openai
langchain-communitys